✨Feat: feed create, update, delete 인증인가 처리 구현

+++ src/main/kotlin/com/sparta/reviewus/domain/feed/service/FeedServiceImpl.kt
@@ -1,5 +1,6 @@
 package com.sparta.reviewus.domain.feed.service
 
+import com.sparta.reviewus.domain.exception.AccessDeniedException
 import com.sparta.reviewus.domain.exception.ModelNotFoundException
 import com.sparta.reviewus.domain.feed.dto.CreateFeedRequest
 import com.sparta.reviewus.domain.feed.dto.FeedByIdResponse
@@ -8,8 +9,8 @@ import com.sparta.reviewus.domain.feed.dto.UpdateFeedRequest
 import com.sparta.reviewus.domain.feed.model.Feed
 import com.sparta.reviewus.domain.feed.model.toResponse
 import com.sparta.reviewus.domain.feed.repository.FeedRepository
+import com.sparta.reviewus.domain.member.dto.AuthenticatedMember
 import com.sparta.reviewus.domain.member.model.toResponse
-import com.sparta.reviewus.domain.member.repository.MemberRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -18,43 +19,23 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class FeedServiceImpl(
     private val feedRepository: FeedRepository,
-    private val memberRepository: MemberRepository,
     private val likeService: LikeService
 ) : FeedService {
-    @Transactional
-    override fun createFeed(request: CreateFeedRequest): FeedResponse {
-        val member = memberRepository.findByIdOrNull(request.memberId) ?: throw ModelNotFoundException(
-            "Member",
-            request.memberId
-        )
-        return feedRepository.save(
-            Feed(
-                title = request.title,
-                category = request.category,
-                feedPicUrls = request.feedPicUrl,
-                longitude = request.longitude,
-                latitude = request.latitude,
-                description = request.description,
-                member = member
-            )
-        ).toResponse()
-    }
 
     override fun getFeeds(): List<FeedResponse> {
         return feedRepository.findAll().map { it.toResponse() }
     }
 
-    override fun getFeedById(id: Long, memberId:Long?): FeedByIdResponse {
-        val feed = feedRepository.findByIdOrNull(id) ?: throw ModelNotFoundException("feed", id)
+    override fun getFeedById(feedId: Long, authenticatedMember: AuthenticatedMember?): FeedByIdResponse {
+        val feed = feedRepository.findByIdOrNull(feedId) ?: throw ModelNotFoundException("feed", feedId)
 
-        val countLikes = likeService.countLikes(id)
+        val countLikes = likeService.countLikes(feedId)
 
-
-        val isLiked = if (memberId != null) likeService.isLiked(memberId, id)
+        val isLiked = if (authenticatedMember != null) likeService.isLiked(authenticatedMember.id, feedId)
         else false
 
         return FeedByIdResponse(
-            id = id!!,
+            id = feedId!!,
             title = feed.title,
             description = feed.description,
             member = feed.member.toResponse(),
@@ -68,17 +49,43 @@ class FeedServiceImpl(
     }
 
     @Transactional
-    override fun deleteFeed(id: Long) {
-        val feed = feedRepository.findByIdOrNull(id) ?: throw ModelNotFoundException("feed", id)
-        feedRepository.delete(feed)
+    override fun createFeed(request: CreateFeedRequest, authenticatedMember: AuthenticatedMember): FeedResponse {
+        return feedRepository.save(
+            Feed(
+                title = request.title,
+                category = request.category,
+                feedPicUrls = request.feedPicUrls,
+                longitude = request.longitude,
+                latitude = request.latitude,
+                description = request.description,
+                member = authenticatedMember.toMember()
+            )
+        ).toResponse()
     }
 
-    override fun updateFeed(id: Long, request: UpdateFeedRequest): FeedResponse {
-        val feed = feedRepository.findByIdOrNull(id) ?: throw ModelNotFoundException("feed", id)
+    @Transactional
+    override fun updateFeed(feedId: Long, request: UpdateFeedRequest, authenticatedMember: AuthenticatedMember): FeedResponse {
+        val feed = feedRepository.findByIdOrNull(feedId) ?: throw ModelNotFoundException("feed", feedId)
+
+        if (feed.member.id != authenticatedMember.id) throw AccessDeniedException()
 
         feed.title = request.title
         feed.description = request.description
+        feed.feedPicUrls = request.feedPicUrls
+        feed.longitude = request.longitude
+        feed.latitude = request.latitude
+        feed.category = request.category
+
         return feedRepository.save(feed).toResponse()
+    }
+
+    @Transactional
+    override fun deleteFeed(feedId: Long, authenticatedMember: AuthenticatedMember) {
+        val feed = feedRepository.findByIdOrNull(feedId) ?: throw ModelNotFoundException("feed", feedId)
+
+        if (feed.member.id != authenticatedMember.id) throw AccessDeniedException()
+
+        feedRepository.delete(feed)
     }
 }
